login/member auth

+++ src/main/java/com/choong/spr/controller/BoardController.java
@@ -1,8 +1,10 @@
 package com.choong.spr.controller;
 
+import java.security.Principal;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -40,7 +42,9 @@ public class BoardController {
 	}
 	
 	@PostMapping("insert")
-	public String insert(BoardDto board, RedirectAttributes rttr) {
+	public String insert(BoardDto board, Principal principal, RedirectAttributes rttr) {
+
+		board.setMemberId(principal.getName());
 		boolean success = service.insertBoard(board);
 		
 		if (success) {
@@ -64,30 +68,44 @@ public class BoardController {
 	}
 	
 	@PostMapping("modify")
-	public String modify(BoardDto dto, RedirectAttributes rttr) {
-		boolean success = service.updateBoard(dto);
+	public String modify(BoardDto dto, Principal principal, RedirectAttributes rttr) {
+		BoardDto oldBoard = service.getBoardById(dto.getId());
 		
-		if (success) {
-			rttr.addFlashAttribute("message", "글이 수정되었습니다.");
+		if (oldBoard.getMemberId().equals(principal.getName())) {
+			boolean success = service.updateBoard(dto);
+			
+			if(success) {
+				rttr.addFlashAttribute("message", "글이 수정되었습니다.");
+			} else {
+				rttr.addFlashAttribute("message", "수정에 실패했습니다.");
+			}
+			
 		} else {
-			rttr.addFlashAttribute("message", "글이 수정되지 않았습니다.");
+			rttr.addFlashAttribute("message", "권한이 없습니다.");
 		}
-		
 		rttr.addAttribute("id", dto.getId());
 		
 		return "redirect:/board/get";
 	}
 	
 	@PostMapping("remove")
-	public String remove(BoardDto dto, RedirectAttributes rttr) {
+	public String remove(BoardDto dto, Principal principal, RedirectAttributes rttr) {
+		BoardDto oldBoard = service.getBoardById(dto.getId());
 		
-		boolean success = service.deleteBoard(dto.getId());
-		
-		if (success) {
-			rttr.addFlashAttribute("message", "글이 삭제 되었습니다.");
+		if(oldBoard.getMemberId().equals(principal.getName())) {
 			
+			boolean success = service.deleteBoard(dto.getId());
+		
+			if (success) {
+				rttr.addFlashAttribute("message", "글이 삭제 되었습니다.");
+			
+			} else {
+				rttr.addFlashAttribute("message", "글이 삭제 되지않았습니다.");
+			}
 		} else {
-			rttr.addFlashAttribute("message", "글이 삭제 되지않았습니다.");
+			rttr.addFlashAttribute("message", "권한이 없습니다.");
+			rttr.addAttribute("id", dto.getId());
+			return "redirect:/board/get";
 		}
 		
 		return "redirect:/board/list";
